+++ src/main/java/nextstep/auth/support/AuthorizationException.java
@@ -1,4 +1,4 @@
-package nextstep.support;
+package nextstep.auth.support;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.ResponseStatus;
